+++ packages/dataviews/src/bulk-actions.tsx
@@ -47,7 +47,7 @@ interface BulkActionsProps< Item > {
 	data: Item[];
 	actions: Action< Item >[];
 	selection: string[];
-	onSelectionChange: SetSelection;
+	onChangeSelection: SetSelection;
 	getItemId: ( item: Item ) => string;
 }
 
@@ -177,7 +177,7 @@ export default function BulkActions< Item >( {
 	data,
 	actions,
 	selection,
-	onSelectionChange,
+	onChangeSelection,
 	getItemId,
 }: BulkActionsProps< Item > ) {
 	const bulkActions = useMemo(
@@ -249,7 +249,7 @@ export default function BulkActions< Item >( {
 						disabled={ areAllSelected }
 						hideOnClick={ false }
 						onClick={ () => {
-							onSelectionChange(
+							onChangeSelection(
 								selectableItems.map( ( item ) =>
 									getItemId( item )
 								)
@@ -263,7 +263,7 @@ export default function BulkActions< Item >( {
 						disabled={ selection.length === 0 }
 						hideOnClick={ false }
 						onClick={ () => {
-							onSelectionChange( [] );
+							onChangeSelection( [] );
 						} }
 					>
 						{ __( 'Deselect' ) }