+++ packages/edit-site/src/components/page-templates/index.js
@@ -232,7 +232,7 @@ export default function PageTemplates() {
 		}
 	);
 	const history = useHistory();
-	const onSelectionChange = useCallback(
+	const onChangeSelection = useCallback(
 		( items ) => {
 			if ( view?.type === LAYOUT_LIST ) {
 				history.push( {
@@ -370,7 +370,7 @@ export default function PageTemplates() {
 				isLoading={ isLoadingData }
 				view={ view }
 				onChangeView={ onChangeView }
-				onSelectionChange={ onSelectionChange }
+				onChangeSelection={ onChangeSelection }
 				selection={ selection }
 				setSelection={ setSelection }
 			/>